requirement.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.3
django-crispy-forms==1.7.2
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
emoji==0.5.1
Faker==1.0.1
httpie==1.0.2
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==5.4.1
pyenchant==2.0.0
pyfiglet==0.8.post0
Pygments==2.3.0
PyJWT==1.7.1
pylint==2.2.2
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
